requirement.txt
python-dotenv
openai
requests
flask
josnify